finalizer implemented works and integrated into wrapper; project finished?!

+++ wrapper.sh
@@ -8,17 +8,34 @@ if [ -f "$infile" ];
 then
 	if [ -f "$accountsFile" ];
 	then
-		./preprocess.sh "$infile" > "pre_""$infile" &&
-		./completifier.py "pre_""$infile" "comp_""$infile" &&
-		rm "pre_""$infile" &&
-		./condenser.py "comp_""$infile" "cond_""$infile" &&
-		rm "comp_""$infile" &&
-		./preprocess.sh "$accountsFile" > "pre_""$accountsFile" &&
-		./makeNumeric.sh "cond_""$infile" "pre_""$accountsFile" "num_""$infile" &&
-		rm "cond_""$infile" && 
-		rm "pre_""$accountsFile" &&
-		rm "script_pre_""$accountsFile"".sh" &&
-		echo "Done" # place holder for next command
+		if [ -f "$outfile" ];
+		then
+			echo "Overwriting the outfile is not supported"
+		else
+			echo "Preprocessing..." &&
+			./preprocess.sh "$infile" > "pre_""$infile" &&
+			echo "    ... done" &&
+			echo "Completing..." &&
+			./completifier.py "pre_""$infile" "comp_""$infile" &&
+			echo "    ... done" &&
+			rm "pre_""$infile" &&
+			echo "Condensing..." &&
+			./condenser.py "comp_""$infile" "cond_""$infile" &&
+			echo "    ... done" &&
+			rm "comp_""$infile" &&
+			echo "Preparing Accountsfile ..." &&
+			./preprocess.sh "$accountsFile" > "pre_""$accountsFile" &&
+			echo "    ... done" &&
+			echo "Making Accounts numeric..." &&
+			./makeNumeric.sh "cond_""$infile" "pre_""$accountsFile" "num_""$infile" &&
+			echo "    ... done" &&
+			rm "cond_""$infile" && 
+			rm "pre_""$accountsFile" &&
+			echo "Finalizing..." &&
+			./finalizer.sh "num_""$infile" "$outfile" &&
+			echo "    ... done" &&
+			rm "num_""$infile"
+		fi
 	else
 		echo "Accounts file does not exist"
 	fi
